requirements.txt
Django==5.1.1
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
factory_boy==3.3.1
mysqlclient==2.2.4
pillow==10.4.0
python-dotenv==1.0.1
